# Basic parfor-loop support in DaphneDSL #515