--- a/build.sh
+++ b/build.sh
@@ -62,6 +62,7 @@ function printHelp {
     echo "  --hdfs            Compile with support for HDFS"
     echo "  --io_uring        Compile with support for io_uring"
     echo "  --no-papi         Compile without support for PAPI"
+    echo "  --enable-parallel-parfor Compile with omp support for parfor statements"
 }
 
 #******************************************************************************
@@ -456,6 +457,7 @@ BUILD_IO_URING="-DUSE_IO_URING=OFF"
 BUILD_PAPI="-DUSE_PAPI=ON"
 WITH_DEPS=1
 WITH_SUBMODULE_UPDATE=1
+enableParallelParfor="OFF"
 
 while [[ $# -gt 0 ]]; do
     key=$1
@@ -527,6 +529,10 @@ while [[ $# -gt 0 ]]; do
         ;;
     -ns | --no-submodule-update)
         WITH_SUBMODULE_UPDATE=0
+        ;;
+    --enable-parallel-parfor)
+        echo enabling parallelization for parfor
+        enableParallelParfor="ON"
         ;;
       *)
         unknown_options="${unknown_options} ${key}"
@@ -1109,6 +1115,7 @@ daphne_msg "Build Daphne"
 
 cmake -S "$projectRoot" -B "$daphneBuildDir" -G Ninja -DANTLR_VERSION="$antlrVersion" \
     -DCMAKE_PREFIX_PATH="$installPrefix" \
+    -DENABLE_PARALLEL_PARFOR="$enableParallelParfor" \
     $BUILD_CUDA $BUILD_FPGAOPENCL $BUILD_DEBUG $BUILD_MPI $BUILD_HDFS $BUILD_PAPI
 
 cmake --build "$daphneBuildDir" --target "$target"