--- a/src/compiler/lowering/LinkParForOutputPass.cpp
+++ b/src/compiler/lowering/LinkParForOutputPass.cpp
@@ -27,7 +27,7 @@
 
 using namespace mlir;
 
-//TODO: REFACTORE ME 
+// TODO: REFACTORE ME
 namespace {
 struct LinkParForOutputPass : public PassWrapper<LinkParForOutputPass, OperationPass<LLVM::LLVMFuncOp>> {
     void runOnOperation() override {
@@ -42,7 +42,7 @@ struct LinkParForOutputPass : public PassWrapper<LinkParForOutputPass, Operation
             return;
 
         // find all GEPOps that have the return arg as their base pointer
-         // output pointer is always passed first for kernels
+        // output pointer is always passed first for kernels
         auto funcOutArg = blocks.front().getArgument(0);
 
         std::vector<LLVM::GEPOp> outGEPOps;
@@ -67,47 +67,94 @@ struct LinkParForOutputPass : public PassWrapper<LinkParForOutputPass, Operation
         OpBuilder b(&getContext());
 
         // rewire outputs of last kernel calls to the respective output of the function.
+        llvm::errs() << "return stores found " << returnStores.size() << "\n";
+        SetVector<Operation *> toErase = {}; 
         for (auto store : returnStores) {
             auto retVal = store->getOperand(0);
             // unrealized cast 1
             auto retValDef = retVal.getDefiningOp();
+            retValDef->dump();
             // unrealized cast 2
-            auto retValDef2 = retValDef->getOperand(0).getDefiningOp();
-            // load
-            auto load = retValDef2->getOperand(0).getDefiningOp();
-            // find the call to last kernel 
-            LLVM::CallOp lastUpdate = nullptr;
-            auto ptr = load->getOperand(0);
-            for (auto usr : ptr.getUsers()) {
-                if ((lastUpdate = llvm::dyn_cast<LLVM::CallOp>(usr))) {
-                    break;
+            mlir::Value operand = retValDef->getOperand(0);
+            // load the output from the kernel
+            operand.dump();
+            if (auto retValDef2 = retValDef->getOperand(0).getDefiningOp()) {
+                auto load = retValDef2->getOperand(0).getDefiningOp();
+                llvm::errs() << "def 0 \n";
+                load->dump();
+                setInPlace(load, store, b, &toErase);
+            } else if (auto blockArg = operand.dyn_cast<mlir::BlockArgument>()) {
+                mlir::Block *parentBlock = blockArg.getOwner();
+                unsigned argIndex = blockArg.getArgNumber();
+
+                for (mlir::Block *pred : parentBlock->getPredecessors()) {
+                    mlir::Operation &terminator = pred->back();
+
+                    if (auto branchInterface = mlir::dyn_cast<mlir::BranchOpInterface>(terminator)) {
+                        auto successors = terminator.getSuccessors();
+
+                        for (unsigned succIdx = 0; succIdx < successors.size(); ++succIdx) {
+                            if (successors[succIdx] == parentBlock) {
+                                auto succOperands = branchInterface.getSuccessorOperands(succIdx);
+                                if (argIndex < succOperands.size()) {
+                                    mlir::Value incoming = succOperands[argIndex];
+                                    llvm::errs() << "def 1 \n";
+                                    incoming.getDefiningOp()->dump();
+                                    setInPlace(incoming.getDefiningOp(), store, b, &toErase);
+                                }
+                            }
+                        }
+                    } else if (terminator.getNumSuccessors() == 1 && terminator.getSuccessor(0) == parentBlock) {
+                        auto succOperands = terminator.getOperands();
+                        if (argIndex < succOperands.size()) {
+                            mlir::Value incoming = succOperands[argIndex];
+                            llvm::errs() << "def 2 \n";
+                            incoming.getDefiningOp()->dump();
+                            setInPlace(incoming.getDefiningOp(), store, b, &toErase);
+                        }
+                    }
                 }
             }
-            // Set the insertion point before lastUpdate
-            b.setInsertionPoint(lastUpdate);
-
-            // Get the defining operation of gep
-            auto gep = store->getOperand(1);
-            auto gepOp = gep.getDefiningOp();
-            auto offset = gepOp->getOperand(1).getDefiningOp();
-
-            // Insert a clone of gepOp at the new location (if moving, use move semantics if supported)
-            auto *clonedGepOp = gepOp->clone();
-            auto *clonedOffset = offset->clone();
-            clonedGepOp->setOperand(1, clonedOffset->getResult(0));
-
-            b.insert(clonedOffset);
-            b.insert(clonedGepOp);
-
-            // Update lastUpdate operand to use the result of the newly inserted gepOp
-            lastUpdate.setOperand(0, clonedGepOp->getResult(0));
-
-            // Erase the old operations
-            store->erase();
-            gepOp->erase();
-           
+        }
+        
+        // erase old operations after rewire 
+        for(auto opToErase : toErase) {
+            opToErase->erase();
         }
         func->removeAttr("parfor_inplace_rewrite_needed");
+    }
+
+    void setInPlace(mlir::Operation *load, mlir::Operation* store, OpBuilder b, SetVector<Operation* > *toErase) {
+        auto ptr = load->getOperand(0);
+        mlir::Operation *lastUpdate = nullptr;
+        for (auto usr : ptr.getUsers()) {
+            if ((lastUpdate = llvm::dyn_cast<LLVM::CallOp>(usr))) {
+                break;
+            }
+        }
+        // Set the insertion point before lastUpdate
+        b.setInsertionPoint(lastUpdate);
+        // Get the defining operation of gep
+        auto gep = store->getOperand(1);
+        auto gepOp = gep.getDefiningOp();
+        auto offset = gepOp->getOperand(1).getDefiningOp();
+
+        // Insert a clone of gepOp at the new location (if moving, use move semantics if supported)
+        auto *clonedGepOp = gepOp->clone();
+        auto *clonedOffset = offset->clone();
+        
+        clonedGepOp->setOperand(1, clonedOffset->getResult(0));
+        
+        b.insert(clonedOffset);
+        b.insert(clonedGepOp);
+        // Update lastUpdate operand to use the result of the newly inserted gepOp
+        lastUpdate->setOperand(0, clonedGepOp->getResult(0));
+        // Load the in-place updated result
+        load->setOperand(0, clonedGepOp->getResult(0));
+        // Erase the old operations
+        toErase->insert(store);
+        toErase->insert(gepOp);
+        toErase->insert(offset);
     }
 };