--- a/src/compiler/lowering/LowerToLLVMPass.cpp
+++ b/src/compiler/lowering/LowerToLLVMPass.cpp
@@ -580,6 +580,7 @@ class ParForOpLowering : public OpConversionPattern<daphne::ParForOp> {
                       
             ++index;
         }
+
         // ********************************************************************
         // Store results in output array as CallKernelOp prescribes. 
         // ********************************************************************