… CMakeLists.txt
… src/compiler/lowering/CMakeLists.txt
… src/compiler/lowering/ManageObjRefsPass.cpp
… src/compiler/lowering/ParForOpLowering.cpp
… src/compiler/lowering/ParForReductionDetectionPass.cpp
… src/compiler/lowering/RewriteToCallKernelOpPass.cpp
… src/compiler/lowering/SpecializeGenericFunctionsPass.cpp
… src/ir/daphneir/DaphneOps.td
… src/ir/daphneir/Passes.h
… src/parser/daphnedsl/DaphneDSLVisitor.cpp
… src/runtime/local/kernels/CMakeLists.txt
… src/runtime/local/kernels/ParFor.h
… src/runtime/local/kernels/kernels.json
… test/api/cli/controlflow/parfor_11.daphne
… test/api/cli/controlflow/parfor_12.daphne
… test/api/cli/controlflow/parfor_13.daphne
… test/api/cli/controlflow/parfor_14.daphne
… test/api/cli/controlflow/parfor_15.daphne
… test/api/cli/controlflow/parfor_16.daphne
… test/api/cli/controlflow/parfor_17.daphne
… test/api/cli/controlflow/parfor_18.daphne
… test/api/cli/controlflow/parfor_19.daphne
… test/api/cli/controlflow/parfor_2.daphne
… test/api/cli/controlflow/parfor_20.daphne
… test/api/cli/controlflow/parfor_3.daphne
… test/api/cli/controlflow/parfor_4.daphne
… test/api/cli/controlflow/parfor_5.daphne
… test/api/cli/controlflow/parfor_6.daphne
… test/api/cli/controlflow/parfor_6.txt
… test/api/cli/controlflow/parfor_7.daphne